working on creating evolution function for each pokemon( WIP)
Figured out I can tie a pokemon to a species and then to their respective evolution trigger in order to get the right id for each pokemon.

Next, I plan to learn how to access the full chain of evolutions to return meaningful data upon each evolution in the nested data in order to properly render data.