Hydrapple now renders on applin's evo chain screen
Added a check for image rendering for complex chains to make sure it grabs the correct index of the github urls for scenarios where the complex split occurs on indexes other than ones it aligns with. For exmaple, silcoon and cascoon will render their images with evo2img[feIndex] but hydrapple will not render with feIndex, so in that case the index switches to seIndex, where it knows its the first of that array.